--- a/src/@types/handshake-interface.ts
+++ b/src/@types/handshake-interface.ts
@@ -1,6 +1,6 @@
+import type { PeerId } from '@libp2p/interfaces/peer-id'
 import type { bytes } from './basic.js'
 import type { NoiseSession } from './handshake.js'
-import type { PeerId } from '@libp2p/interfaces/peer-id'
 
 export interface IHandshake {
   session: NoiseSession