--- a/src/index.ts
+++ b/src/index.ts
@@ -1,4 +1,7 @@
 import { Noise } from './noise.js'
+
+export * from './crypto.js'
+export * from './crypto/stablelib.js'
 export * from './noise.js'
 
 /**