--- a/test/keycache.spec.ts
+++ b/test/keycache.spec.ts
@@ -1,9 +1,9 @@
+import type { PeerId } from '@libp2p/interfaces/peer-id'
+import { Buffer } from 'buffer'
 import { assert } from 'chai'
+import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
 import { KeyCache } from '../src/keycache.js'
 import { createPeerIds, createPeerIdsFromFixtures } from './fixtures/peer.js'
-import { equals as uint8ArrayEquals } from 'uint8arrays/equals'
-import { Buffer } from 'buffer'
-import type { PeerId } from '@libp2p/interfaces/peer-id'
 
 describe('KeyCache', () => {
   let peerA: PeerId