Hello.
This PR extracts all low level crypto interface in a specific file and makes the crypto interface pluggable in `Noise`.
Hope this helps!