@mpetrunic 2022-03-24
> Looks good to me, nice abstraction! Are there any performance tests to ensure equal throughput? Results should be the same, but just to be sure

master
$ node benchmarks/benchmark.js
Initializing handshake benchmark
Init complete, running benchmark
handshake x 40.53 ops/sec ±17.26% (72 runs sampled)
Done in 6.54s.

PR#133
$ node benchmarks/benchmark.js
Initializing handshake benchmark
Init complete, running benchmark
handshake x 41.62 ops/sec ±15.59% (72 runs sampled)
Done in 6.39s.

It varies on each run but it's identical