@mpetrunic 2022-03-24
@dapplion This potentially allows you in lodestar to use some native library that might be faster than stablelib (purejs).