@dapplion 2022-03-25
> @dapplion This potentially allows you in lodestar to use some native library that might be faster than stablelib (purejs).

yeah looking forward to testing multiple impl