@ShogunPanda 2022-03-25
If you want, I have sample implementation with sodium-native in https://github.com/web3-storage/ipfs-elastic-provider-bitswap-peer/blob/main/src/noise-crypto.js