@dapplion 2022-04-03
> If you want, I have sample implementation with sodium-native in https://github.com/web3-storage/ipfs-elastic-provider-bitswap-peer/blob/main/src/noise-crypto.js

Let's goo! :fire: I you wanna contribute bringing this live in Lodestar I may airdrop some $LODE tokens in the future