@MarcoPolo 2023-03-29
I've implemented this interface with WASM using Rust and the `snow` crate: https://github.com/MarcoPolo/js-libp2p-noise-wasm

Here's some benchmarks:

Throughput:
```
❮ node dist/benchmark/src/index.js
TCP+mplex+noise
Download throughput is (mbits/s) 369n
Upload throughput is (mbits/s) 272n
TCP+yamux+noise
Download throughput is (mbits/s) 193n
Upload throughput is (mbits/s) 195n
TCP+mplex+wasm_noise
Download throughput is (mbits/s) 1047n
Upload throughput is (mbits/s) 1055n
TCP+yamux+wasm_noise
Download throughput is (mbits/s) 869n
Upload throughput is (mbits/s) 867n
```

Handshakes per second:

```
❮ node benchmarks/benchmark.js
Initializing handshake benchmark
Init complete, running benchmark
handshake x 632 ops/sec ±0.58% (89 runs sampled)
```